requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.18.5
Django==3.1
ffmpy==0.2.3
Flask==1.1.2
Flask-ApiExceptions==1.1.2
gunicorn==20.1.0
httpx==0.13.3
itsdangerous==1.1.0
Jinja2==2.11.2
matplotlib==3.2.2
MarkupSafe==1.1.1
librosa==0.7.2
numba==0.48
pathos==0.2.6
phonemizer~=2.2.1
p_tqdm==1.3.3
pydub==0.24.1
pyworld==0.3.0
protobuf==3.20.0
ruamel.yaml>=0.16.6
scipy==1.4.1
soundfile==0.10.3.post1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorboard==2.3.0
tensorflow-addons==0.17.1
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
webrtcvad==2.0.10
Werkzeug==1.0.1
nltk==3.5
PyYAML==5.3.1
git+https://github.com/zxp54332/g2p.git
streamlit==1.8.1
gTTS==2.2.2
googletrans==3.1.0a0
pandas==1.4.0